+++ underworld/systems/_stokes.py
@@ -86,16 +86,8 @@ class Stokes(_stgermain.StgCompoundComponent):
     _selfObjectName = "_system"
 
     def __init__(self, velocityField, pressureField, fn_viscosity, fn_bodyforce=None, fn_lambda=None, voronoi_swarm=None, conditions=[],
-                _removeBCs=True, _fn_viscosity2=None, _fn_director=None, _fn_stresshistory=None, swarm=None, **kwargs):
+                _removeBCs=True, _fn_viscosity2=None, _fn_director=None, _fn_stresshistory=None, **kwargs):
 
-        # DEPRECATE. JM 09/16
-        if swarm:
-            import warnings
-            warnings.warn("'swarm' paramater has been renamed to 'voronoi_swarm'. Please update your models. "+
-                          "'swarm' parameter will be removed in the next release.")
-            if voronoi_swarm:
-                raise ValueError("Please provide only a 'voronoi_swarm'. 'swarm' is deprecated.")
-            voronoi_swarm = swarm
         if not isinstance( velocityField, uw.mesh.MeshVariable):
             raise TypeError( "Provided 'velocityField' must be of 'MeshVariable' class." )
         if velocityField.nodeDofCount != velocityField.mesh.dim: